Loaders for docx and pptx
