+++ ragondin/components/indexer/loaders/CustomDocLoader.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from langchain_core.documents.base import Document
 from langchain_community.document_loaders import UnstructuredWordDocumentLoader, UnstructuredODTLoader
 
-from components.indexer.loaders.BaseLoader import BaseLoader
+from .base import BaseLoader
 
 
 class CustomDocLoader(BaseLoader):
