+++ ragondin/components/indexer/loaders/Custompymupdf4llm.py
@@ -1,7 +1,7 @@
 import pymupdf4llm
 
 from langchain_core.documents.base import Document
-from components.indexer.loaders.BaseLoader import BaseLoader
+from .base import BaseLoader
 
 class Custompymupdf4llm(BaseLoader):
     def __init__(self, page_sep: str='[PAGE_SEP]', config=None, **kwargs) -> None:
