+++ ragondin/components/indexer/loaders/VideoAudioLoader.py
@@ -2,16 +2,14 @@ import gc
 import os
 import torch
 import whisperx
-
-
 from pathlib import Path
 from loguru import logger
 from pydub import AudioSegment
 
 from langchain_core.documents.base import Document
 
-from components.indexer.loaders.BaseLoader import BaseLoader
-from components.indexer.loaders.AudioTranscriber import AudioTranscriber
+from .base import BaseLoader
+from .AudioTranscriber import AudioTranscriber
 
 class VideoAudioLoader(BaseLoader):
     def __init__(self, page_sep: str='[PAGE_SEP]', batch_size=4, config=None):
